Added context cc for ancestors, and (l,r,ll,rr,lll,rrr,llr,rrl) for left/right limits. Doesn't work for parents yet.
git-svn-id: svn+ssh://beta.visl.sdu.dk/usr/local/svn/repos/visl/tools/vislcg3/trunk@8368 cb2587b3-b7ff-0310-8c81-a2a651690ada

+++ src/GrammarApplicator_matchSet.cpp
@@ -771,7 +771,7 @@ bool GrammarApplicator::doesSetMatchCohortNormal_helper(ReadingList& readings, c
 	return false;
 }
 
-bool GrammarApplicator::doesSetMatchCohortNormal(Cohort& cohort, const uint32_t set, const ContextualTest *test, uint32_t options) {
+bool GrammarApplicator::doesSetMatchCohortNormal(Cohort& cohort, const uint32_t set, const ContextualTest *test, uint64_t options) {
 	/*
 	return !doesSetMatchCohort(cohort, set, options).empty();
 	/*/
@@ -814,7 +814,7 @@ bool GrammarApplicator::doesSetMatchCohortCareful_helper(ReadingList& readings, 
 	return !readings.empty();
 }
 
-bool GrammarApplicator::doesSetMatchCohortCareful(Cohort& cohort, const uint32_t set, const ContextualTest *test, uint32_t options) {
+bool GrammarApplicator::doesSetMatchCohortCareful(Cohort& cohort, const uint32_t set, const ContextualTest *test, uint64_t options) {
 	/*
 	return !doesSetMatchCohort(cohort, set, options).empty();
 	/*/
